refactor: `wrap-root-element` api를 컴포넌트로 분리 (#15)
`wrap-root-element`가 컴포넌트 타입이 아니기 때문에
내부에서 hooks을 사용할 수 없는 문제를 해결하고자
`wrap-root-element.ts`에서 담당하던 로직을 `wrapper.ts`로
이동하고 `wrap-root-element.ts`에서는 `wrapper` 컴포넌트를
사용하여 element를 렌더하도록 함
